fix(spentcalories): remove unnecessary type conversion in distance()

+++ internal/spentcalories/spentcalories.go
@@ -53,7 +53,7 @@ func parseTraining(data string) (int, string, time.Duration, error) {
 }
 
 func distance(steps int, height float64) float64 {
-	return height * stepLengthCoefficient * float64(steps) / float64(mInKm)
+	return height * stepLengthCoefficient * float64(steps) / mInKm
 }
 
 func meanSpeed(steps int, height float64, duration time.Duration) float64 {
